requirements.txt
pandas
numpy
yfinance
requests
python-dotenv
